Added memories and calendar screens (have to incorporate into reports screen still), added new logos to all pages, and generateDemoData
The memories and calendar screens were added.
The new logo was added to all screens.
generateDemoData (populates data for a child for a whole year) is available in scripts folder. To run, edit childID and userID to fit the relevant data, and in terminal, navigate to project directory and run command: "run generate-demo."